GOOGLE-115
jsTree v3.0.0 was released, but it contained a bug that caused it to not work when nodes contained ul-elements other than its own.  I (lsloan) reported this on the jsTree forum and the developer fixed it in the trunk of the project.  This version of jsTree is that unreleased update, which fixed this bug.  This version should be replaced with the future release that includes this fix.