GOOGLE-153
Added more event types to handle item background shading for searches and item deletions.  Moved the code to a new function for cleaner-looking code.  Limited application of shaded background to only visible items, to be compatible with search results.  Deferred the shading update to later in the JS queue so that it happens after the delete operation has completed.